fix(feature): fix followed feed updating incorrectly
Assume that:
1. User is already subscribed to subreddit 'A'
2. Starts from the Home screen
3. User's selected feed is 'Followed'

If user subscribes to 'B', returns to the Home screen, and switches to the Feed screen, the followed feed will display posts from both 'A' and 'B'.

To fix this, change behaviour of onIconButtonClick in HomeScreen such that item is deleted from the database the moment the lambda fires off. Item is inserted back into the database when undo button on snackbar is pressed.

 Swap AnimatedVisibility implementation with Animatable to preserve animations.