fix(feature): stop re-navigation when launching intents after cold boot
When app is launched via an intent, configuration changes cause the app to 're-navigate' to the intent's associated destination. This happens because the intent resolver is fired off every time a configuration change occurs.

Surround the intent resolver with an additional check to verify if the app has already resolved the intent to prevent this.