feat(core): rework download service
Add back getHLSPlaylist method, which no longer returns the result, but rather saves it into the hlsVideoUrl and hlsAudioUrl variables.

Split downloadHLS into downloadVideo & downloadAudio.

Add optional videoUrl parameter in downloadVideo. If it is empty, default to value in hlsVideoUrl. This is because some videos on Redlib do not have a HLS variant.

Add fileName parameter to set the downloaded file name.