refactor(core): add id parameter to DownloadData
Intention is for the Id parameter to be used instead of hash codes of file names for download ids. For now, we will use the 'id' portion of the pathname of the download URL for this purpose.

e.g., The 'v1zT4' part of https://example.com/vid/v1zT4/1080.mp4